requirements.txt
# 💥 Erro de Preço Bot — Dependências
python-telegram-bot==20.7
APScheduler==3.10.4
aiohttp==3.9.1
beautifulsoup4==4.12.3
requests==2.31.0
flask==3.0.0
lxml==5.1.0
